% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_afs.R
\name{theme_afs}
\alias{theme_afs}
\title{theme_afs
Creates AFS themed ggplot2 theme}
\usage{
theme_afs(font_family)
}
\arguments{
\item{font_family}{Choose which font you want to use: "Classic", "Modern".
Default is "Modern"}
}
\value{
ggplot object
}
\description{
theme_afs
Creates AFS themed ggplot2 theme
}
\examples{
ggplot(cars, aes(speed, dist)) +
geom_point(alpha = 0.75) +
labs(title = "Cars Example Theme", subtitle = "Subtitle example",
caption = "Caption Example\nMore lines here"
) +
theme_afs()
}
